requirement.txt
absl-py==2.1.0
attrs==23.2.0
blinker==1.7.0
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.3
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cvzone==1.6.1
cycler==0.12.1
dlib==19.24.2
encode==0.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.13.1
Flask==3.0.2
Flask-Login==0.6.3
flatbuffers==23.5.26
fonttools==4.49.0
fsspec==2024.2.0
gmpy2==2.1.2
h11==0.8.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
idna==2.10
imutils==0.5.4
itsdangerous==2.1.2
jax==0.4.25
Jinja2==3.1.3
jwt==1.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.3
mediapipe==0.10.10
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
mysql-connector-python==8.3.0
networkx==3.1
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pip==24.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
scipy==1.12.0
seaborn==0.13.2
setuptools==68.2.2
six==1.16.0
sounddevice==0.4.6
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.17.1+cu118
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.20
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2
win-inet-pton==1.1.0
